Create dedicated module for DiscussionCourseXBlock.
