+++ openedx/core/lib/xblock_builtin/xblock_discussion/__init__.py
@@ -1,1 +1,2 @@
-from .xblock_discussion import DiscussionXBlock, DiscussionCourseXBlock
+from .xblock_discussion import DiscussionXBlock
+from .xblock_discussion_course import DiscussionCourseXBlock
