+++ openedx/core/lib/xblock_builtin/xblock_discussion/setup.py
@@ -29,7 +29,7 @@ setup(
     entry_points={
         'xblock.v1': [
             'discussion = xblock_discussion:DiscussionXBlock',
-            'discussion-course = xblock_discussion:DiscussionCourseXBlock',
+            'discussion-course = xblock_discussion_course:DiscussionCourseXBlock',
         ]
     },
     package_data=package_data("xblock_discussion", ["static"]),
