+++ openedx/core/lib/xblock_builtin/xblock_discussion/utils.py
@@ -1,4 +1,16 @@
+"""
+Utils for Discussion XBlock and Course Discussion XBlock
+"""
+
 from django.templatetags.static import static
+
+
+def _(text):
+    """
+    A noop underscore function that marks strings for extraction.
+    """
+    return text
+
 
 def asset_to_static_url(asset_path):
     """
